create:DB接続メソッドの作成（ファイル名リネーム）
